apps/api/pyproject.toml
[project]
name = "wikigit-api"
version = "0.1.0"
description = "WikiGit FastAPI backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "gitpython>=3.1.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pydantic-settings-yaml>=0.2.0",
    "whoosh>=2.7.4",
    "python-frontmatter>=1.1.0",
    "aiofiles>=24.1.0",
    "python-multipart>=0.0.17",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
